Totalmente funcional mas ainda faltam meios de editar/deletar/excluir os Tipos de Entradas, Tipos de Rendas e Despesas mais comuns e um glitch no menu de navegacao para Registro que somente devera aparecer se o usuario ainda nao tiver sido registrado. E por fim, traduzir para duas linguas.

+++ kakeibo.api/Models/Despesas.cs
@@ -11,11 +11,11 @@ namespace kakeibo.api
         [DatabaseGenerated(DatabaseGeneratedOption.Identity)]
         [Required(ErrorMessage = "Despesa ID is required")]
         [Display(Name = "Despesa ID")]
-        public required int DespesaID { get; set; } // int, not null
+        public required decimal DespesaID { get; set; } // int, not null
 
         [Required(ErrorMessage = "Tipo Despesa ID is required")]
         [Display(Name = "Tipo Despesa ID")]
-        public required int TipoDespesaID { get; set; } // int, not null
+        public required decimal TipoDespesaID { get; set; } // int, not null
 
         [MaxLength(80)]
         [StringLength(80)]
